kafka_config: Utility function for formatting configuration variables
The Kafka module classes need the configuration in various formats. For
example, some need a hash describing the brokers, some need just the
list of their FQDNs, while others need a concatenated string of
host:port pairs. Starting from a cluster name, a hash of all of the
brokers in all of the clusters and a hash or list of zookeeper hosts,
the function returns one hash conveniently keyed and formatted ready to
be used.

Bug: T130371
Change-Id: I64dcd3cceaed749bbae6e88fdf588b79ddf49fa5